Edited generate code function
Removed Generate Code function. Added download and upload.
This is basically a function that allows the user to download text and upload it into a new program to restore all words.
A function to generate delaminated text that contains contents and a function to parse delaminated text and upload it to library have been added.